# perf: add a generic CID type which allows for aliased byte slices