@MichaelMure 2025-01-21
Imho this is valuable and should be revisited. Code is still available at https://github.com/ipfs/go-cid/compare/468cca0d172b911fa0a8a0f59ba087682531fea6